(#57) Introduce wall and plane modules
